+++ app/src/main/res/layout/activity_login.xml
@@ -13,4 +13,18 @@
         android:layout_height="match_parent"
         />
 
+    <include
+        layout="@layout/include_error_layout"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:layout_below="@id/toolbar"
+        />
+
+    <include
+        layout="@layout/include_progress_layout"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:layout_below="@id/toolbar"
+        />
+
 </RelativeLayout>
